+++ orm/encoding/ormkv/seq.go
@@ -4,7 +4,7 @@ import (
 	"bytes"
 	"encoding/binary"
 
-	"github.com/cosmos/cosmos-sdk/orm/types/ormerrors"
+	"cosmossdk.io/orm/types/ormerrors"
 
 	"google.golang.org/protobuf/reflect/protoreflect"
 )