+++ orm/encoding/ormkv/seq_test.go
@@ -4,12 +4,12 @@ import (
 	"bytes"
 	"testing"
 
-	"github.com/cosmos/cosmos-sdk/orm/encoding/ormkv"
+	"cosmossdk.io/orm/encoding/ormkv"
 
 	"gotest.tools/v3/assert"
 	"pgregory.net/rapid"
 
-	"github.com/cosmos/cosmos-sdk/orm/internal/testpb"
+	"cosmossdk.io/orm/internal/testpb"
 )
 
 func TestSeqCodec(t *testing.T) {