+++ orm/internal/stablejson/encode_test.go
@@ -10,8 +10,8 @@ import (
 	bankv1beta1 "cosmossdk.io/api/cosmos/bank/v1beta1"
 	basev1beta1 "cosmossdk.io/api/cosmos/base/v1beta1"
 	txv1beta1 "cosmossdk.io/api/cosmos/tx/v1beta1"
+	"cosmossdk.io/orm/internal/stablejson"
 	"github.com/cosmos/cosmos-proto/anyutil"
-	"github.com/cosmos/cosmos-sdk/orm/internal/stablejson"
 )
 
 func TestStableJSON(t *testing.T) {