+++ orm/model/ormtable/index_impl.go
@@ -3,18 +3,18 @@ package ormtable
 import (
 	"context"
 
-	"github.com/cosmos/cosmos-sdk/orm/types/kv"
+	"cosmossdk.io/orm/types/kv"
 
-	"github.com/cosmos/cosmos-sdk/orm/internal/fieldnames"
+	"cosmossdk.io/orm/internal/fieldnames"
 
-	"github.com/cosmos/cosmos-sdk/orm/model/ormlist"
+	"cosmossdk.io/orm/model/ormlist"
 
-	"github.com/cosmos/cosmos-sdk/orm/types/ormerrors"
+	"cosmossdk.io/orm/types/ormerrors"
 
 	"google.golang.org/protobuf/proto"
 	"google.golang.org/protobuf/reflect/protoreflect"
 
-	"github.com/cosmos/cosmos-sdk/orm/encoding/ormkv"
+	"cosmossdk.io/orm/encoding/ormkv"
 )
 
 // indexKeyIndex implements Index for a regular IndexKey.